Setup low-fidelity initial page a user would see if they were to access "postogon.com" without already being signed in.
IonButtons have an attribute called routerLink and routerDirection.
Since this is expected to be the root of our application, we will ensure that each routerDirection has a "forward" animation.

IonButton 1: Sign in with email
routerLink='/signin'

IonButton 2: New account
routerLink='/signup'

IonButton 3: Trouble getting in?
routerLink='/recovery'

We are working on a low fidelity design for now which will be enhanced in the future. The purpose is to have a robust application flow. We can go back and focus on the design of the UI at a later time.

We will have three more buttons, but they won't be using IonButton. We will use the regular html button tag and style it later through TailwindCSS. These will open a modal when clicked. (onClick).

We will create a single separate component to handle this modal because otherwise we would be rewriting the same code 3 times.

Button 1: Terms of Service
Button 2: Privacy Policy
Button 3: Cookies Policy

For summary, so far we created one page that has 3 buttons that are intended to take us to three page that we intend to create next. We have 3 other buttons that open up a modal which we also intend to create.

Up next: Creating 3 pages and a modal.

Created Recovery.tsx, Signin.tsx, and signup.tsx

Signin.tsx and signup.tsx have an ion item with a floating label n an inpute for both an email and a password.

Next step is to setup the recovery page.