Created Reusable Policy Component, introduced Constants and enforced typing.  Added more space to margins in home.tsx, made recovery button actually take you to the recovery page.
In the future, revisit policycontainer.tsx within the components folder and have it take in the constant policy.<policyname> directly, from home.tsx. Only reason I hesitated because we can now use use this container for other cases outside of holding our policy information. That is why we should revisit this.